backend/requirements.txt
httpx
fastapi
python-dotenv
uvicorn
langchain-text-splitters
tiktoken
sentence-transformers
qdrant-client
openai
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary